requirements.txt
jupyter
numpy
matplotlib
scikit-image
numpy-stl